+++ libs/core/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@hicommonwealth/schemas": "workspace:*",
+    "@hicommonwealth/evm-protocols": "workspace:*",
     "dotenv": "^16.0.3",
     "ethers": "5.7.2",
     "lodash": "^4.17.21",
