+++ libs/core/src/integration/chain-event.utils.ts
@@ -1,4 +1,8 @@
 import { EventNames, EventPairs } from '@hicommonwealth/core';
+import {
+  EvmEventSignature,
+  EvmEventSignatures,
+} from '@hicommonwealth/evm-protocols';
 import { ETHERS_BIG_NUMBER, EVM_ADDRESS } from '@hicommonwealth/schemas';
 import { BigNumber } from 'ethers';
 import { decodeLog } from 'web3-eth-abi';
@@ -8,38 +12,10 @@ import { z } from 'zod';
  * To add a new chain-event:
  * 1. Add the event signature to EvmEventSignatures
  * 2. Create a Zod schema for the Common equivalent of the parsed event and add it to ChainEventSchemas
- * 3.Add the Event inputs ABI to EvmEventAbis
+ * 3. Add the Event inputs ABI to EvmEventAbis
  * 4. Create a mapping function that implements the EventMapper type
  * 5. Add the new mapper to the EvmMappers object
  */
-
-export const EvmEventSignatures = {
-  NamespaceFactory: {
-    NamespaceDeployed:
-      '0x8870ba2202802ce285ce6bead5ac915b6dc2d35c8a9d6f96fa56de9de12829d5',
-    ContestManagerDeployed:
-      '0x990f533044dbc89b838acde9cd2c72c400999871cf8f792d731edcae15ead693',
-  },
-  Contests: {
-    ContentAdded:
-      '0x2f0d66b98c7708890a982e2194479b066a117a6f9a8f418f7f14c6001965b78b',
-    RecurringContestStarted:
-      '0x32391ebd47fc736bb885d21a45d95c3da80aef6987aa90a5c6e747e9bc755bc9',
-    RecurringContestVoterVoted:
-      '0x68d40dd5e34d499a209946f8e381c1258bdeff6dea4e96e9ab921da385c03667',
-    SingleContestStarted:
-      '0x002817006cf5e3f9ac0de6817ca39830ac7e731a4949a59e4ac3c8bef988b20c',
-    SingleContestVoterVoted:
-      '0xba2ce2b4fab99c4186fd3e0a8e93ffb61e332d0c4709bd01d01e7ac60631437a',
-  },
-  CommunityStake: {
-    Trade: '0xfc13c9a8a9a619ac78b803aecb26abdd009182411d51a986090f82519d88a89e',
-  },
-} as const;
-
-type Values<T> = T[keyof T];
-type NestedValues<T> = Values<{ [K in keyof T]: Values<T[K]> }>;
-export type EvmEventSignature = NestedValues<typeof EvmEventSignatures>;
 
 // Event Inputs can be found in the contract ABI by filtering the objects by type = 'event' e.g.:
 //   for (const obj of abi) {
