+++ libs/core/tsconfig.build.json
@@ -5,5 +5,12 @@
     "outDir": "build"
   },
   "include": ["src"],
-  "references": [{ "path": "../schemas/tsconfig.build.json" }]
+  "references": [
+    {
+      "path": "../schemas/tsconfig.build.json"
+    },
+    {
+      "path": "../evm-protocols/tsconfig.build.json"
+    }
+  ]
 }
