--- libs/model/src/chainEventSignatures.ts
@@ -1,17 +1,0 @@
-export const deployedNamespaceEventSignature =
-  '0x8870ba2202802ce285ce6bead5ac915b6dc2d35c8a9d6f96fa56de9de12829d5';
-
-export const communityStakeTradeEventSignature =
-  '0xfc13c9a8a9a619ac78b803aecb26abdd009182411d51a986090f82519d88a89e';
-
-export const launchpadTokenLaunchedEventSignature =
-  '0xd7ca5dc2f8c6bb37c3a4de2a81499b25f8ca8bbb3082010244fe747077d0f6cc';
-
-export const launchpadTradeEventSignature =
-  '0x9adcf0ad0cda63c4d50f26a48925cf6405df27d422a39c456b5f03f661c82982';
-
-export const launchpadTokenRegisteredEventSignature =
-  '0xc2fe88a1a3c1957424571593960b97f158a519d0aa4cef9e13a247c64f1f4c35';
-
-export const communityNamespaceCreatedEventSignature =
-  '0xa16d784cb6c784b621c7877ce80495765ed32ca0b3dba2ef467116a435f125fd';
