+++ libs/model/src/index.ts
@@ -30,7 +30,6 @@ export type { E2E_TestEntities } from './tester';
 export * as middleware from './middleware';
 
 // Internals - Should not be exported once we finish the migrations to models
-export * from './chainEventSignatures';
 export * from './config';
 export * from './database';
 export * from './models';
