+++ libs/model/src/services/commonProtocol/newNamespaceValidator.ts
@@ -1,8 +1,6 @@
 import { AppError, ServerError } from '@hicommonwealth/core';
-import {
-  communityNamespaceCreatedEventSignature,
-  models,
-} from '@hicommonwealth/model';
+import { communityNamespaceCreatedEventSignature } from '@hicommonwealth/evm-protocols';
+import { models } from '@hicommonwealth/model';
 import { BalanceSourceType, commonProtocol } from '@hicommonwealth/shared';
 import Web3 from 'web3';
 import { CommunityAttributes } from '../../models';
